requirements.txt
clip==0.2.0
einops==0.8.0
h5py==3.10.0
imageio==2.33.1
kornia==0.7.2
labels==20.1.0
matplotlib==3.4.3
more_itertools==10.2.0
natsort==8.4.0
numpy==1.21.2
omegaconf==2.3.0
packaging==21.0
pandas==2.0.3
Pillow==9.5.0
pudb==2024.1
pytorch_lightning==1.4.2
PyYAML==6.0.1
Requests==2.31.0
scikit_learn==1.3.0
setuptools==52.0.0
torch_geometric==2.0.1
tqdm==4.61.2
transformers==4.40.2
typing_extensions==3.10.0.2
wandb==0.16.6
torchmetrics==0.7.0
test-tube
monai==0.9.1
timm==1.0.15